fix(raids): remove intel grade import cycle
